+++ src/Capsule.php
@@ -4,9 +4,10 @@ declare(strict_types=1);
 
 namespace PhpCfdi\XmlCancelacion;
 
+use Countable;
 use DateTimeImmutable;
 
-class Capsule
+class Capsule implements Countable
 {
     /** @var string */
     private $rfc;
@@ -14,7 +15,7 @@ class Capsule
     /** @var DateTimeImmutable */
     private $date;
 
-    /** @var array */
+    /** @var array This is a B-Tree array, values are stored in keys */
     private $uuids;
 
     public function __construct(string $rfc, array $uuids = [], DateTimeImmutable $date = null)
@@ -22,7 +23,7 @@ class Capsule
         $this->rfc = $rfc;
         $this->date = $date ?? new DateTimeImmutable('now');
         foreach ($uuids as $uuid) {
-            $this->append($uuid);
+            $this->uuids[strtoupper($uuid)] = true;
         }
     }
 
@@ -39,11 +40,6 @@ class Capsule
     public function uuids(): array
     {
         return array_keys($this->uuids);
-    }
-
-    public function append(string $uuid): void
-    {
-        $this->uuids[strtoupper($uuid)] = true;
     }
 
     public function count(): int